+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Using the same concept for SLJsonSerializer as for SLJsonDeserializer
 - SLJsonNode.Parse changed to parse not only objects, but also arrays and values
 - SLJsonConverter changed to generic delegate comparable with Func<T, TResult> to keep compatible with .NET 2.0 and 3.0
+
 ## [1.0.0] - 2019-05-13
 ### Added
 - CHANGELOG.md introduced